Add "analyzer_token_limit", limits stream for hit_source anlayzer
This is in an attempt to limit damage when we hit a large document
that isn't using postings or fvh. We simply close off the stream
when we hit this limit. I'm assuming the TokenStream is lazy in
for analysis so we can avoid analyzing the whole source if we hit
the token limit before the end.

An alternative we can use if this doesn't help with performance is
just to truncate the `source` directly which will surely avoid
analysis although then potentially introduces artefacts on the
truncation boundary, unless we do some more intelligent boundary
detection, but then I guess we're back to analysis!